change to synchronized(ly)

+++ src/base/multithreading/unsynch/TransferRunnable.java
@@ -18,7 +18,7 @@ public class TransferRunnable implements Runnable {
 		try {
 			while (true) {
 				int toAccount = (int)(bank.size() * Math.random());
-				double amount = maxAmount * Math.random();
+				int amount = (int)(maxAmount * Math.random());
 				bank.transfer(fromAccount, toAccount, amount);
 				Thread.sleep((long) (DELAY * Math.random()));
 			}
